#1 refactor : POJO 클래스들을 main 패키지로 리팩토링 한다
