+++ src/main/java/snowflake/tumblog/user/domain/repository/UserAdapter.java
@@ -1,0 +1,17 @@
+package snowflake.tumblog.user.domain.repository;
+
+import snowflake.tumblog.user.domain.User;
+
+public class UserAdapter implements UserPort {
+
+    private final UserRepository userRepository;
+
+    public UserAdapter(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public void save(User user) {
+        userRepository.save(user);
+    }
+}
