+++ src/main/java/snowflake/tumblog/user/domain/repository/UserPort.java
@@ -1,0 +1,8 @@
+package snowflake.tumblog.user.domain.repository;
+
+import snowflake.tumblog.user.domain.User;
+
+public interface UserPort {
+
+    void save(User user);
+}
