+++ src/main/java/snowflake/tumblog/user/dto/CreateUserRequest.java
@@ -1,0 +1,10 @@
+package snowflake.tumblog.user.dto;
+
+import org.springframework.util.Assert;
+
+public record CreateUserRequest(String nickname) {
+
+    public CreateUserRequest {
+        Assert.notNull(nickname, "nickname must not be null");
+    }
+}
